requirements.txt
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.10.3
visitor==0.1.3
Werkzeug==0.15.5
WTForms==2.2.1
